Remove rust code that isnt used anymore
