+++ ironfish-rust-nodejs/src/lib.rs
@@ -50,36 +50,6 @@ pub fn generate_new_public_address(private_key: String) -> Result<Key> {
     })
 }
 
-#[napi(object)]
-pub struct MineHeaderNapiResult {
-    pub randomness: f64,
-    pub found_match: bool,
-}
-
-#[napi]
-pub fn mine_header_batch(
-    mut header_bytes: Buffer,
-    initial_randomness: i64,
-    target_buffer: Buffer,
-    batch_size: i64,
-) -> MineHeaderNapiResult {
-    let mut target_array = [0u8; 32];
-    target_array.copy_from_slice(&target_buffer[..32]);
-
-    // Execute batch mine operation
-    let mine_header_result = mining::mine_header_batch(
-        header_bytes.as_mut(),
-        initial_randomness,
-        &target_array,
-        batch_size,
-    );
-
-    MineHeaderNapiResult {
-        randomness: mine_header_result.randomness,
-        found_match: mine_header_result.found_match,
-    }
-}
-
 #[napi(constructor)]
 pub struct FoundBlockResult {
     pub randomness: f64,
